Fixed build and setup state machine to follow moving ground
