+++ Assets/Samples/SimplifiedDemoKCC/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bfe86701af5a70840b1fe7fc8931ff8f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
