+++ Assets/Samples/SimplifiedDemoKCC/Editor/DrawCameraController.cs.meta
