+++ Assets/Samples/SimplifiedDemoKCC/Editor/nickmaltbie.OpenKCC.Demo 1.asmdef
@@ -1,0 +1,21 @@
+{
+    "name": "nickmaltbie.OpenKCC.Demo.Editor",
+    "rootNamespace": "com.nickmaltbie.OpenKCC.Demo.Editor",
+    "references": [
+        "nickmaltbie.OpenKCC",
+        "nickmaltbie.OpenKCC.Editor",
+        "Unity.InputSystem",
+        "nickmaltbie.StateMachineUnity"
+    ],
+    "includePlatforms": [
+        "Editor"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
