+++ Assets/Samples/SimplifiedDemoKCC/Editor/nickmaltbie.OpenKCC.Demo 1.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8c4d03566481193498ba43c8aa7731ca
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
