+++ Assets/StreamingAssets/UnityServicesProjectConfiguration.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 12eafa9c27e533f4cae359c012063765
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
